Tag: V00-07-08 2010-08-31 Andy Salnikov - O2OHdf5Writer - use file creation property list to change the sizes   of the b-rtrees in a file, this should reduce overhead that we see   for XPP data
git-svn-id: https://pswww.slac.stanford.edu/svn-readonly/psdmrepo@1173 b967ad99-d558-0410-b138-e0f6c56caec7

+++ O2OTranslator/trunk/include/DgramReader.h
@@ -62,7 +62,7 @@ public:
                 size_t maxDgSize,
                 O2OXtcMerger::MergeMode mode,
                 bool skipDamaged,
-                int l1OffsetSec = 0 ) ;
+                double l1OffsetSec = 0 ) ;
 
   // Destructor
   ~DgramReader () ;
@@ -80,7 +80,7 @@ private:
   size_t m_maxDgSize ;
   O2OXtcMerger::MergeMode m_mode ;
   bool m_skipDamaged ;
-  int m_l1OffsetSec ;
+  double m_l1OffsetSec ;
 
 };
